+++ src/main/java/com/ll/medium/MediumApplication.java
@@ -1,4 +1,4 @@
-package com.ll.medium.medium;
+package com.ll.medium;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
