+++ src/main/java/com/ll/medium/domain/member/memeber/controller/MemberController.java
@@ -1,0 +1,16 @@
+package com.ll.medium.domain.member.memeber.controller;
+
+import lombok.RequiredArgsConstructor;
+import org.springframework.stereotype.Controller;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.RestController;
+
+@Controller
+@RestController("/member")
+@RequiredArgsConstructor
+public class MemberController {
+    @GetMapping("/join")
+    public String showJoin() {
+        return "domain/member/member/join";
+    }
+}
