+++ src/main/resources/application-test.yml
@@ -1,13 +1,12 @@
 spring:
-  output:
-    ansi:
-      enabled: always
   datasource:
     url: jdbc:h2:mem:medium_test;MODE=MYSQL
   jpa:
     hibernate:
       ddl-auto: create
 custom:
+  site:
+    baseUrl: "http://localhost:8080"
   tempDirPath: c:/temp/medium_test/temp
   genFile:
     dirPath: c:/temp/medium_test
