+++ src/main/resources/templates/domain/member/member/join.html
@@ -1,0 +1,3 @@
+<h1>회원가입</h1>
+
+<form action=""></form>
