+++ src/test/java/com/ll/medium/domain/home/home/controller/HomeControllerTest.java
@@ -1,0 +1,22 @@
+package com.ll.medium.domain.home.home.controller;
+
+import org.junit.jupiter.api.DisplayName;
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.autoconfigure.web.servlet.AutoConfigureMockMvc;
+import org.springframework.boot.test.context.SpringBootTest;
+import org.springframework.test.context.ActiveProfiles;
+import org.springframework.transaction.annotation.Transactional;
+
+@SpringBootTest
+@AutoConfigureMockMvc
+@ActiveProfiles("test")
+@Transactional
+class HomeControllerTest  {
+
+    @DisplayName("t1")
+    @Test
+    void t1() {
+
+    }
+
+}
