--- src/test/java/com/ll/medium/domain/home/home/controller/MediumApplicationTests.java
@@ -1,18 +1,0 @@
-package com.ll.medium.medium;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-@AutoConfigureMockMvc
-@ActiveProfiles("test")
-@Transactional
-class HomeControllerTest  {
-
-    @DisplayName("t1")
-    @Test
-    void t1() {
-
-    }
-
-}
